# web/openapi.json
{
  "components": {
    "schemas": {
      "AddRequest": {
        "properties": {
          "external_key": {
            "type": "string"
          },
          "source_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": ["source_id", "external_key"],
        "type": "object"
      },
...
      "ChapterDto": {
        "properties": {
          "external_key": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "language": {
            "type": ["string", "null"]
          },
          "manga_id": {
            "format": "uuid",
            "type": "string"
          },
          "number": {
            "format": "float",
            "type": ["number", "null"]
          },
          "published_at": {
            "format": "date-time",
            "type": ["string", "null"]
          },
          "title": {
            "type": ["string", "null"]
          },
          "volume": {
            "format": "float",
            "type": ["number", "null"]
          }
        },
        "required": ["id", "manga_id", "external_key"],
        "type": "object"
      },
...
      "MangaDto": {
        "properties": {
          "artists": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "authors": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "content_rating": {
            "type": "string"
          },
          "cover_url": {
            "type": ["string", "null"]
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": ["string", "null"]
          },
          "external_key": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "language": {
            "type": ["string", "null"]
          },
          "reading_direction": {
            "type": "string"
          },
          "source_id": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "url": {
            "type": ["string", "null"]
          }
        },
        "required": [
          "id",
          "source_id",
          "external_key",
          "title",
          "authors",
          "artists",
          "tags",
          "status",
          "content_rating",
          "reading_direction",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
      "Paged_ChapterDto": {
        "description": "One page of results.\n\n`next_cursor` is opaque and is the only supported way to ask for the next\npage. Offsets are deliberately absent: they shift under inserts, and a\nlibrary that gains a chapter mid-scroll would skip or repeat one.",
        "properties": {
          "items": {
            "items": {
              "properties": {
                "external_key": {
                  "type": "string"
                },
                "id": {
                  "format": "uuid",
                  "type": "string"
                },
                "language": {
                  "type": ["string", "null"]
                },
                "manga_id": {
                  "format": "uuid",
                  "type": "string"
                },
                "number": {
                  "format": "float",
                  "type": ["number", "null"]
                },
                "published_at": {
                  "format": "date-time",
                  "type": ["string", "null"]
                },
                "title": {
                  "type": ["string", "null"]
                },
                "volume": {
                  "format": "float",
                  "type": ["number", "null"]
                }
              },
              "required": ["id", "manga_id", "external_key"],
              "type": "object"
            },
            "type": "array"
          },
          "next_cursor": {
            "type": ["string", "null"]
          }
        },
        "required": ["items"],
        "type": "object"
      },
...
      "Paged_MangaDto": {
        "description": "One page of results.\n\n`next_cursor` is opaque and is the only supported way to ask for the next\npage. Offsets are deliberately absent: they shift under inserts, and a\nlibrary that gains a chapter mid-scroll would skip or repeat one.",
        "properties": {
          "items": {
            "items": {
              "properties": {
                "artists": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "authors": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "content_rating": {
                  "type": "string"
                },
                "cover_url": {
                  "type": ["string", "null"]
                },
                "created_at": {
                  "format": "date-time",
                  "type": "string"
                },
                "description": {
                  "type": ["string", "null"]
                },
                "external_key": {
                  "type": "string"
                },
                "id": {
                  "format": "uuid",
                  "type": "string"
                },
                "language": {
                  "type": ["string", "null"]
                },
                "reading_direction": {
                  "type": "string"
                },
                "source_id": {
                  "format": "uuid",
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "tags": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "title": {
                  "type": "string"
                },
                "updated_at": {
                  "format": "date-time",
                  "type": "string"
                },
                "url": {
                  "type": ["string", "null"]
                }
              },
              "required": [
                "id",
                "source_id",
                "external_key",
                "title",
                "authors",
                "artists",
                "tags",
                "status",
                "content_rating",
                "reading_direction",
                "created_at",
                "updated_at"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "next_cursor": {
            "type": ["string", "null"]
          }
        },
        "required": ["items"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session": {
        "description": "An opaque, signed, HttpOnly session cookie issued by `GET /api/v1/auth/callback`. There is no bearer token: tokens from the identity provider never reach the browser.",
        "in": "cookie",
        "name": "id",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "A self-hosted manga library server.",
    "license": {
      "name": "AGPL-3.0-or-later"
    },
    "title": "nyuka",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/chapters/{id}": {
      "get": {
        "operationId": "getChapter",
        "parameters": [
          {
            "description": "Chapter id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "No such chapter"
          }
        },
        "summary": "`GET /api/v1/chapters/{id}`",
        "tags": ["library"]
      }
    },
...
    "/manga": {
      "get": {
        "operationId": "listLibrary",
        "parameters": [
          {
            "description": "The `next_cursor` from a previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paged_MangaDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "`GET /api/v1/manga`",
        "tags": ["library"]
      },
      "post": {
        "description": "The body names a source and a key, not a series: the metadata comes from\nthe source, never from the client. Accepting a client's version would let\nanyone write any title into the library.",
        "operationId": "addMangaToLibrary",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MangaDto"
                }
              }
            },
            "description": "Added, or already present"
          },
          "404": {
            "description": "No such source or series"
          },
          "502": {
            "description": "The source could not be reached"
          }
        },
        "summary": "`POST /api/v1/manga` — add a catalog entry to the library.",
        "tags": ["catalog"]
      }
    },
    "/manga/{id}": {
      "get": {
        "operationId": "getManga",
        "parameters": [
          {
            "description": "Series id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MangaDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "No such series"
          }
        },
        "summary": "`GET /api/v1/manga/{id}`",
        "tags": ["library"]
      }
    },
    "/manga/{id}/chapters": {
      "get": {
        "operationId": "listMangaChapters",
        "parameters": [
          {
            "description": "Series id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The `next_cursor` from a previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paged_ChapterDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "No such series"
          }
        },
        "summary": "`GET /api/v1/manga/{id}/chapters`",
        "tags": ["library"]
      }
    },
...
  },
...
}